Parking Lot Mapping System
Uses a Raspberry Pi Model 3B wih the PiCamera module to send images of a parking lot over a network to a host for processing.  Use a base image to select stalls and let program run to output an image depicting if the stall is open. The program will evaluate the RGB values of each stall and will compare the stalls value to the base.